Updated player movement, added max ball speed, and stopped death on collision with wall
