+++ OpenGL/Ball.cpp
@@ -43,7 +43,10 @@ void Ball::Render()
 
 void Ball::OnCollisionEnter(Pawn* _other)
 {
-	m_physicsBody->SetLinearVelocity(b2Vec2(2 * m_physicsBody->GetLinearVelocity().x + 1, 2 * m_physicsBody->GetLinearVelocity().y + 1));
+	while (m_physicsBody->GetLinearVelocity().Length() < 10)
+	{
+		m_physicsBody->SetLinearVelocity(b2Vec2(m_physicsBody->GetLinearVelocity().x + 1, m_physicsBody->GetLinearVelocity().y + 1));
+	}
 }
 
 void Ball::checkgate(b2Vec2 _gate, int &playerscore) {
