+++ OpenGL/OpenGL.vcxproj.filters
@@ -87,6 +87,9 @@
     <ClCompile Include="PlayerShield.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Input.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\Shaders\3D.vs">
