+++ OpenGL/Player.h
@@ -17,5 +17,7 @@ public:
 	void Die();
 
 private:
+	float m_fMoveSpeed = 10.0f;
+
 	CPlayerShield m_shield;
 };
