+++ OpenGL/Player2.h
@@ -20,5 +20,7 @@ public:
 	void Respawn();
 
 private:
+	float m_fMoveSpeed = 10.0f;
+
 	CPlayerShield m_shield;
 };
