+++ OpenGL/Utilities.h
@@ -8,21 +8,6 @@
 const int WINDOW_WIDTH = 800;
 const int WINDOW_HEIGHT = 600;
 
-enum InputState
-{
-	INPUT_FIRST_RELEASE,
-	INPUT_RELEASED,
-	INPUT_FIRST_PRESS,
-	INPUT_HOLD
-};
-
-enum InputMouse
-{
-	MOUSE_LEFT,
-	MOUSE_MIDDLE,
-	MOUSE_RIGHT
-};
-
 enum GameState
 {
 	MENU,
@@ -30,21 +15,4 @@ enum GameState
 	GAMEOVER
 };
 
-
-void Keyboard_Down(unsigned char key, int x, int y);
-
-void Keyboard_Up(unsigned char key, int x, int y);
-
-void Mouse(int _iButton, int _iGlutState, int _iX, int _iY);
-
-bool GetButtonDown(unsigned char _cKey);
-bool GetButtonUp(unsigned char _cKey);
-
-bool GetMouseButtonDown(int _iButton);
-bool GetMouseButtonUp(int _iButton);
-
-float GetDeltaRotation();
-
 void RotateVector(b2Vec2& _vector, float _fDeg);
-
-bool MenuButtons();
